@@ -9,13 +9,14 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.7,<=3.11"
+python = ">=3.9,<=3.11"
 web3 = "^5.4.0"
 ccxt = "^3.1.15"
 
 [tool.poetry.group.dev.dependencies]
 autonomy-dev = {extras = ["all"], version = "^0.1.4"}
 jupyterlab = "^4.0.1"
+types-requests = "^2.31.0.1"
 
 [build-system]
 requires = ["poetry-core"]
