@@ -1,10 +1,15 @@
+"""
+This module contains the RyskOptionMarket class.
+"""
 from dataclasses import dataclass
 from datetime import datetime
+
+from rysk_client.src.utils import get_contract
 
 
 @dataclass
 class RyskOptionMarket:
-    """"""
+    """Rysk option market."""
 
     strike: float
     expiration: int
@@ -35,9 +40,7 @@ class RyskOptionMarket:
         return self._parse_name()
 
     def __str__(self):
-        return "RyskOptionMarket(name={} strike={}, expiration={}, is_put={})".format(
-            self.name, self.strike, self.expiration, self.is_put
-        )
+        return f"RyskOptionMarket({self.name})"
 
 
 class RyskOptionMarketManager:
@@ -46,7 +49,7 @@ class RyskOptionMarketManager:
     def fetch_option_markets(self):
         """Fetches the option markets from the API"""
         markets = []
-        for expiration in expirations:
+        for expiration in self._fetch_expirations():
             put_strikes = self.option_catalogue.functions.getOptionDetails(
                 expiration, True
             ).call()
@@ -55,45 +58,24 @@ class RyskOptionMarketManager:
             ).call()
 
             for strike in put_strikes:
-                markets.append(RyskMarket(expiration, strike, True))
+                markets.append(RyskOptionMarket(expiration, strike, True))
 
             for strike in call_strikes:
-                markets.append(RyskMarket(expiration, strike, False))
-        markets
+                markets.append(RyskOptionMarket(expiration, strike, False))
+        self.option_markets = markets
+        return markets
 
     def __init__(self, web3):
         self.web3 = web3
         self.option_markets = {}
-        self.fetch_option_markets()
         self.option_catalogue = get_contract("option_catalogue", web3)
+        self.fetch_option_markets()
 
     def fetch_option_market(self, name):
         """
         Fetches a specific option market from the smart contract.
         """
 
-
-# tests/test_rysk_option_market_manager.py
-import pytest
-
-from rysk_client.src.rysk_option_market import RyskOptionMarket, RyskOptionMarketManager
-from rysk_client.src.utils import get_web3
-
-
-@pytest.fixture
-def rysk_option_market_manager():
-    """Returns a RyskOptionMarketManager"""
-
-    web3 = get_web3()
-    return RyskOptionMarketManager(web3)
-
-
-def test_rysk_option_market_manager(rysk_option_market_manager):
-    """Tests the RyskOptionMarketManager class"""
-    assert isinstance(rysk_option_market_manager, RyskOptionMarketManager)
-
-
-def test_fetch_option_markets(rysk_option_market_manager):
-    """Tests the fetch_option_markets method"""
-    rysk_option_market_manager.fetch_option_markets()
-    assert len(rysk_option_market_manager.option_markets) > 0
+    def _fetch_expirations(self):
+        """Returns the expirations from the smart contract."""
+        return self.option_catalogue.functions.getExpirations().call()
